+++ lib/bake/options/options.rb
@@ -19,17 +19,19 @@ module Bake
     attr_accessor :build_config, :nocache, :analyze, :eclipseOrder, :envToolchain, :showConfigs
     attr_reader :main_dir, :project, :filename, :main_project_name, :buildDirDelimiter, :dot, :cc2j_filename # String
     attr_reader :roots, :include_filter, :exclude_filter, :adapt # String List
-    attr_reader :file_list, :conversion_info, :stopOnFirstError, :clean, :rebuild, :show_includes, :show_includes_and_defines, :projectPaths, :qac # Boolean
+    attr_reader :conversion_info, :stopOnFirstError, :clean, :rebuild, :show_includes, :show_includes_and_defines, :projectPaths, :qac # Boolean
     attr_reader :linkOnly, :compileOnly, :no_autodir, :clobber, :docu, :debug, :prepro, :oldLinkOrder, :prebuild, :printTime, :json, :wparse # Boolean
     attr_reader :threads, :socket # Fixnum
     attr_reader :vars # map
     attr_reader :verbose
+    attr_reader :filelist # set
     attr_reader :consoleOutput_fullnames, :consoleOutput_visualStudio
 
 
     def initialize(argv)
       super(argv)
 
+      @filelist = nil
       @qac = false
       @projectPaths = false
       @wparse = false
@@ -39,7 +41,6 @@ module Bake
       @buildDirDelimiter = "/"
       @oldLinkOrder = false
       @conversion_info = false
-      @file_list = false
       @envToolchain = false
       @analyze = false
       @eclipseOrder = false
@@ -92,7 +93,8 @@ module Bake
 
       add_option(["--create"                                     ], lambda { |x| Bake::Create.proj(x)                    })
       add_option(["--conversion-info",    "--conversion_info"    ], lambda {     @conversion_info = true                 })
-      add_option(["--file-list",          "--file_list"          ], lambda {     @file_list = true                       })
+      add_option(["--file-list=json",     "--file_list=json"     ], lambda {     @filelist = Set.new; @json=true         })
+      add_option(["--file-list=txt",      "--file_list=txt"      ], lambda {     @filelist = Set.new                     })
       add_option(["--filter-paths"                               ], lambda {     @projectPaths = true                    })
       add_option(["--qac"                                        ], lambda {     @qac = true                             })
 
@@ -159,8 +161,8 @@ module Bake
         end
       end
 
-      if @conversion_info || @file_list
-        if @conversion_info && @file_list
+      if @conversion_info || @filelist
+        if @conversion_info && @filelist
           Bake.formatter.printError("Error: --conversion-info and --file-list not allowed at the same time")
           ExitHelper.exit(1)
         end