+++ spec/testdata/fileList/main/src/main/main1.cpp
@@ -1,3 +1,8 @@
 #include "main2.h"
 #include "sub1.h"
 #include "main1.h"
+
+int main()
+{
+	return 0;
+}